Nest current ingestions in the chat window
